+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog for ceph3us-features branch fork of Rollbar for Android 
 
-**0.1.3-bc**
+**0.1.3bc1**
 - initial release.
-- add allow to customize Rollbar thread name to be more friendly and  better recognizable as default one 
-
+- add allow to customize Rollbar thread name to be more friendly and  better recognizable as default one
+ 
+**0.1.3bc4**
+- add predefined levels eg: Rollbar.Level.DEBUG
+- rewrite RollbarThread include: 
+    - concurrency logic first step to minimize deadlocking  
+    - more verbosity for debugging